Move empty week summary to dayCard.tsx. Move map methods to the availabilityCalculatorService.
